+++ src/Components/BidderPortfolio/BidderPortfolioPage/BidderPortfolioLoader.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import createLoader from '../../Loadable';
+
+export const path = () => import('./BidderPortfolioPage');
+
+const BidderPortfolio = createLoader({ path, shouldPreload: false });
+
+const BidderPortfolioLoadable = ({ ...rest }) => (
+  <BidderPortfolio {...rest} />
+);
+
+export default BidderPortfolioLoadable;
