+++ src/Components/GlossaryEditor/GlossaryEditorPage/GlossaryEditorLoadable.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import createLoader from '../../Loadable';
+
+export const path = () => import('./GlossaryEditorPage');
+
+const GlossaryEditorPage = createLoader({ path, shouldPreload: false });
+
+const GlossaryEditorPageLoadable = ({ ...rest }) => (
+  <GlossaryEditorPage {...rest} />
+);
+
+export default GlossaryEditorPageLoadable;
